- refactor kategori barang

+++ app/Http/Controllers/Admin/MasterData/KategoriBarangController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin\MasterData;
 
 use Illuminate\Support\Str;
+use Illuminate\Support\Arr;
 use Illuminate\Http\Request;
 use App\Models\BarangKategori;
 use App\Http\Controllers\Controller;
@@ -100,31 +101,39 @@ class KategoriBarangController extends Controller
 
 		$request->validate([
 			'nama'        => 'required|string|max:50',
-			'is_dropdown' => 'nullable|in:ya,tidak'
+			'is_dropdown' => 'required|in:ya,tidak'
 		], [], [
 			'nama'        => 'Kategori Barang',
 			'is_dropdown' => 'Dropdown E-Commerce'
 		]);
 
+		$arr = $request->only('nama');
+
 		$slug        = Str::slug($request->nama);
 		$slug_exists = BarangKategori::where('slug', $slug)->exists();
 
 		if ($slug != $data->slug && $slug_exists) {
-
 			return back()->withErrors(['nama' => 'Kategori barang sudah tersedia.'])->withInput();
 		} else if ($slug != $data && !$slug_exists) {
-
-			$dropdown = $request->is_dropdown == 'ya' ? 1 : 0;
-			$data->update([
-				'nama'        => $request->nama,
-				'slug'        => $slug,
-				'is_dropdown' => $dropdown
-			]);
+			$arr = Arr::add($arr, 'slug', $slug);
 		}
-		alert()
-			->success('Data berhasil diubah', 'Sukses!')
-			->persistent('Tutup')
-			->autoclose(3000);
+
+		$count_dropdown = BarangKategori::where('is_dropdown', 1)->count();
+		if ($count_dropdown == 3 && $request->is_dropdown == 'ya') {
+			alert()
+				->error('Maksimal hanya boleh 3 kategori barang pada menu dropdown.', 'Gagal!')
+				->persistent('Tutup');
+		} else {
+			$dropdown = $request->is_dropdown == 'ya' ? 1 : 0;
+			$arr = Arr::add($arr, 'is_dropdown', $dropdown);
+
+			$data->update($arr);
+			alert()
+				->success('Data berhasil diubah', 'Sukses!')
+				->persistent('Tutup')
+				->autoclose(3000);
+		}
+
 
 		return redirect()->route('admin.master-data.kategori-barang.edit', $uuid);
 	}
@@ -135,14 +144,20 @@ class KategoriBarangController extends Controller
 	 * @param  int  $id
 	 * @return \Illuminate\Http\Response
 	 */
-	public function destroy($uuid)
+	public function destroy(BarangKategori $data)
 	{
-		$data = BarangKategori::findOrFail($uuid);
-		$data->delete();
-		alert()
-			->success('Data berhasil dihapus', 'Sukses!')
-			->persistent('Tutup')
-			->autoclose(3000);
+		if ($data->Barang()->count()) {
+			alert()
+				->error('Terdapat barang pada kategori ini.<br/>Harap ubah kategori pada barang terkait terlebih dahulu.', 'Gagal!')
+				->html()
+				->persistent('Tutup');
+		} else {
+			$data->delete();
+			alert()
+				->success('Data berhasil dihapus', 'Sukses!')
+				->persistent('Tutup')
+				->autoclose(3000);
+		}
 
 		return redirect()->route('admin.master-data.kategori-barang.index');
 	}
